feat(NodeType): Roadiz now generates a new Doctrine Migration for each updates on node-types and node-types fields.
BREAKING CHANGE: `bin/console themes:migrate` command do not execute Doctrine migrations, generate NS entities and schema update. **You must version your NS\*\*\*\*.php files and migrations** to sync your app in different environments.

- `bin/console themes:migrate` and `bin/console themes:install` are deprecated: use `bin/console app:install` to just import data in database or `bin/console app:migrate` to import data and generate Entities and Doctrine migrations

+++ lib/Rozier/src/Controllers/NodeTypes/NodeTypesController.php
@@ -12,6 +12,7 @@ use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Themes\Rozier\Forms\NodeTypeType;
@@ -22,10 +23,12 @@ use Twig\Error\RuntimeError;
 class NodeTypesController extends RozierApp
 {
     private MessageBusInterface $messageBus;
+    private KernelInterface $kernel;
 
-    public function __construct(MessageBusInterface $messageBus)
+    public function __construct(KernelInterface $kernel, MessageBusInterface $messageBus)
     {
         $this->messageBus = $messageBus;
+        $this->kernel = $kernel;
     }
 
     public function indexAction(Request $request): Response
@@ -56,11 +59,8 @@ class NodeTypesController extends RozierApp
     }
 
     /**
-     * Return an edition form for requested node-type.
-     *
      * @param Request $request
      * @param int $nodeTypeId
-     *
      * @return Response
      * @throws RuntimeError
      */
@@ -81,7 +81,6 @@ class NodeTypesController extends RozierApp
         if ($form->isSubmitted() && $form->isValid()) {
             try {
                 $this->em()->flush();
-
                 $this->messageBus->dispatch(new Envelope(new UpdateNodeTypeSchemaMessage($nodeType->getId())));
 
                 $msg = $this->getTranslator()->trans('nodeType.%name%.updated', ['%name%' => $nodeType->getName()]);
@@ -102,8 +101,6 @@ class NodeTypesController extends RozierApp
     }
 
     /**
-     * Return an creation form for requested node-type.
-     *
      * @param Request $request
      *
      * @return Response
@@ -114,24 +111,30 @@ class NodeTypesController extends RozierApp
         $this->denyAccessUnlessGranted('ROLE_ACCESS_NODETYPES');
         $nodeType = new NodeType();
 
-        $form = $this->createForm(NodeTypeType::class, $nodeType);
+        $form = $this->createForm(NodeTypeType::class, $nodeType, [
+            'disabled' => !$this->kernel->isDebug()
+        ]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            try {
-                $this->em()->persist($nodeType);
-                $this->em()->flush();
+            if (!$this->kernel->isDebug()) {
+                $form->addError(new FormError('You cannot create a node-type in production mode.'));
+            } else {
+                try {
+                    $this->em()->persist($nodeType);
+                    $this->em()->flush();
 
-                $this->messageBus->dispatch(new Envelope(new UpdateNodeTypeSchemaMessage($nodeType->getId())));
+                    $this->messageBus->dispatch(new Envelope(new UpdateNodeTypeSchemaMessage($nodeType->getId())));
 
-                $msg = $this->getTranslator()->trans('nodeType.%name%.created', ['%name%' => $nodeType->getName()]);
-                $this->publishConfirmMessage($request, $msg, $nodeType);
+                    $msg = $this->getTranslator()->trans('nodeType.%name%.created', ['%name%' => $nodeType->getName()]);
+                    $this->publishConfirmMessage($request, $msg, $nodeType);
 
-                return $this->redirectToRoute('nodeTypesEditPage', [
-                    'nodeTypeId' => $nodeType->getId()
-                ]);
-            } catch (EntityAlreadyExistsException $e) {
-                $form->addError(new FormError($e->getMessage()));
+                    return $this->redirectToRoute('nodeTypesEditPage', [
+                        'nodeTypeId' => $nodeType->getId()
+                    ]);
+                } catch (EntityAlreadyExistsException $e) {
+                    $form->addError(new FormError($e->getMessage()));
+                }
             }
         }
 
@@ -163,12 +166,16 @@ class NodeTypesController extends RozierApp
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->messageBus->dispatch(new Envelope(new DeleteNodeTypeMessage($nodeType->getId())));
+            if (!$this->kernel->isDebug()) {
+                $form->addError(new FormError('You cannot delete a node-type in production mode.'));
+            } else {
+                $this->messageBus->dispatch(new Envelope(new DeleteNodeTypeMessage($nodeType->getId())));
 
-            $msg = $this->getTranslator()->trans('nodeType.%name%.deleted', ['%name%' => $nodeType->getName()]);
-            $this->publishConfirmMessage($request, $msg, $nodeType);
+                $msg = $this->getTranslator()->trans('nodeType.%name%.deleted', ['%name%' => $nodeType->getName()]);
+                $this->publishConfirmMessage($request, $msg, $nodeType);
 
-            return $this->redirectToRoute('nodeTypesHomePage');
+                return $this->redirectToRoute('nodeTypesHomePage');
+            }
         }
 
         $this->assignation['form'] = $form->createView();
